[BUGFIX][IO_FILE] if directory not exist, it wont try to delete stuff

+++ Assets/Scripts/HamiIO/HamiIoFile.cs
@@ -6,6 +6,7 @@ namespace HamiIO
     {
         public static void DeleteAllFiles(string folder)
         {
+            if(!Directory.Exists(CONSTS.__FULL_PATH_TO_RESOURCES + folder)) return;
             DirectoryInfo di = new DirectoryInfo(CONSTS.__FULL_PATH_TO_RESOURCES + folder);
             foreach (FileInfo file in di.GetFiles()) file.Delete();
         }
